S104 split: extract terraform_ops.py
main.py: 2,293 -> 1,728 lines.
